moving test util function
